pyproject.toml
[tool.poetry]
name = "prototype-auto-cloud"
version = "0.1.0"
description = ""
authors = ["Гальцов Сергей <s.galtsov@dssl.ru>"]
readme = "README.md"
packages = [{include = "prototype_auto_cloud"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.2"
tox = "^4.4.7"
selene = {version = "^2.0.0b17", allow-prereleases = true}
allure-pytest = "^2.13.1"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
